# static/swagger.yaml
swagger: "2.0"
info:
  title: Password Keeper
  version: 1.0.0
  description: >-
    API for managing users, roles, folders, passwords, password history, audit logs,
    and security questions. 
    

    All protected endpoints require a bearer token in the Authorization header. Use the "/user/login" endpoint to obtain a token, then include it as "Bearer <token>" for authenticated requests.


    How to use this API:
      1. Create a user with /user/.
      2. Sign in with /user/login to receive an auth token.
      3. Copy the token and add it to the Authorization header as "Bearer <token>".
      4. Call protected endpoints such as /folder/, /password/, /roles/, or /security/.
      5. If your token expires, use /user/refresh with a refresh token to obtain a new one.
host: 127.0.0.1:5000
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
responses:
  BadRequestError:
    description: Malformed JSON or an invalid request body. The API may return a validation object for schema problems.
    schema:
      $ref: '#/definitions/Error'
  UnauthorizedError:
    description: Missing, expired, or invalid bearer token.
    schema:
      $ref: '#/definitions/Error'
  ForbiddenError:
    description: User is authenticated but lacks the required role.
    schema:
      $ref: '#/definitions/Error'
  NotFoundError:
    description: Requested resource was not found.
    schema:
      $ref: '#/definitions/Error'
  ConflictError:
    description: The request conflicts with an existing resource or current state.
    schema:
      $ref: '#/definitions/Error'
  UnprocessableEntityError:
    description: Business rule validation failed. The response uses the standard error envelope with error_code and details.
    schema:
      $ref: '#/definitions/Error'
  ServerError:
    description: The server encountered an unexpected error while processing the request.
    schema:
      $ref: '#/definitions/Error'
paths:
...
  /roles/:
    get:
      tags: [Roles]
      summary: List roles
      description: Returns all available roles. This endpoint is typically used by administrators and other privileged clients.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Roles returned
          schema:
            type: array
            items:
              $ref: '#/definitions/Role'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
    post:
      tags: [Roles]
      summary: Create a role
      description: Creates a new role. Admin access is required for this operation.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/RoleCredentials'
      responses:
        '201':
          description: Role created
          schema:
            $ref: '#/definitions/Role'
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '409':
          $ref: '#/responses/ConflictError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
    put:
      tags: [Roles]
      summary: Update a role
      description: Updates an existing role and propagates the role change to the affected users.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '201':
          description: Role updated
          schema:
            $ref: '#/definitions/Role'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '409':
          $ref: '#/responses/ConflictError'
        '500':
          $ref: '#/responses/ServerError'
    delete:
      tags: [Roles]
      summary: Delete a role
      description: Deletes a role that is not protected. Admin access is required, and protected roles cannot be removed.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '200':
          description: Role deleted
          schema:
            $ref: '#/definitions/DeleteResponse'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '403':
          $ref: '#/responses/ForbiddenError'
        '404':
          $ref: '#/responses/NotFoundError'
        '409':
          $ref: '#/responses/ConflictError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
...
  /security/:
    get:
      tags: [Security Question]
      summary: List security questions
      description: Returns the authenticated user's security questions and their stored answers metadata.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Security questions returned
          schema:
            type: array
            items:
              $ref: '#/definitions/SecurityQuestion'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
    post:
      tags: [Security Question]
      summary: Create a security question
      description: Stores a new security question for the authenticated user. Use this when adding a new recovery prompt.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SecurityQuestionCredentials'
      responses:
        '201':
          description: Security question created
          schema:
            $ref: '#/definitions/SecurityQuestion'
          examples:
            application/json:
              question_id: 3
              user_id: 1
              question: What is your favorite pet?
              encripted_answer: 1a2b3c4d5e6f7g8h
        '401':
          $ref: '#/responses/UnauthorizedError'
        '409':
          $ref: '#/responses/ConflictError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
    put:
      tags: [Security Question]
      summary: Update a security question
      description: Updates an existing security question for the authenticated user. Send the updated question payload in the request body.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SecurityQuestion'
      responses:
        '201':
          description: Security question updated
          schema:
            $ref: '#/definitions/SecurityQuestion'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
    delete:
      tags: [Security Question]
      summary: Delete a security question
      description: Deletes a security question for the authenticated user. Use this when removing a recovery prompt.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SecurityQuestion'
      responses:
        '200':
          description: Security question deleted
          schema:
            $ref: '#/definitions/DeleteResponse'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/ServerError'
definitions:
...
  Role:
    type: object
    required: [role_name]
    example:
      role_id: 2
      role_name: user
    properties:
      role_id:
        type: integer
      role_name:
        type: string
  RoleCredentials:
    type: object
    required: [role_name]
    example:
      role_name: editor
    properties:
      role_name:
        type: string
...
  SecurityQuestion:
    type: object
    required: [user_id, question, encripted_answer]
    properties:
      question_id:
        type: integer
      user_id:
        type: integer
      question:
        type: string
      encripted_answer:
        type: string
  SecurityQuestionCredentials:
    type: object
    required: [user_id, question, encripted_answer]
    properties:
      user_id:
        type: integer
      question:
        type: string
      encripted_answer:
        type: string
...
  DeleteResponse:
    type: object
    example:
      message: Resource removed successfully
    properties:
      message:
        type: string
  Error:
    type: object
    example:
      status: error
      message: Password not found
      error_code: password_not_found
      details:
        code: password_not_found
        domain: password
        operation: lookup
        message: The requested password could not be found.
    properties:
      status:
        type: string
        description: Always "error" for failed responses.
      message:
        type: string
        description: Human-readable explanation of the failure.
      error_code:
        type: string
        description: Stable machine-readable code for the error category.
      details:
        type: object
        description: Structured information describing the domain, operation, and detail message for the failure.
        properties:
          code:
            type: string
          domain:
            type: string
          operation:
            type: string
          message:
            type: string